refactor(api): Move robot_configs.py to config
