+++ api/src/opentrons/deck_calibration/endpoints.py
@@ -2,7 +2,7 @@ from aiohttp import web
 from uuid import uuid1
 from opentrons.legacy_api.instruments import pipette_config
 from opentrons import instruments, robot
-from opentrons.legacy_api.robot import robot_configs
+from opentrons.config import robot_configs
 from opentrons.deck_calibration import jog, position, dots_set, z_pos
 from opentrons.deck_calibration.linal import add_z, solve
 from typing import Dict, Tuple
